Row/column permutations on separate files.
Permutations of matrix row and columns may be specified on external files now,
in addition to those specified in the matrix definition files. The reason for
this is that different PCB designs for the same keyboard may require different
rearrangements, so the permutations are actually PCB-, not keyboard-specific.
Use option -p to give a file containing the permutations.

Former option -p (don't put generated definitions into RAM) has now -P.