Debouncing can be disabled per firmware image.
If a keyboard is equipped with non-bouncing switches, then the number of
debounce iterations (DEBOUNCE_ITERATIONS) may now be set to 1 to entirely
disable software debouncing.

Also, the default number of iterations for debouncing (if DEBOUNCE_ITERATIONS
is undefined) has been lowered from 10 to 5 as I didn't experience bouncing
effects with that value. When setting the number of iterations to 3, however, I
had occasional double key activations, so 5 seems fine for now.