Partial builds if some tools or libs are missing.
The configure script does not error out anymore if, i.e., the cross compiler or
Python are not found; instead, a warning is printed. If a certain tool or
library is missing that is required to build a some part of the package, then
no attempt will be made to build that part. For instance, if the cross compiler
was not found, then the firmware images cannot be built, but this does not
prevent us from building the boot loader command line tool.

In addition, all warnings issued by configure, if any, will be repeated at the
end, after the configure script has done its main tasks.