Using libusb 1.0 for the new tool.
Turns out that libusb 0.1.x and 1.0.x can be used together with no problems, so
we are ditching 0.1 for the Keyboard Upgrade command line tool, and use it only
for building the boot loader tool. Compared to 0.1.x, library version 1.0.x is
a dream. :)

The configure script checks for both versions. As before, the Makefiles build
only those programs that can actually be built.