Changed layout of stored key maps.
There is a fixed number of characters for the key map name: 20. In previous
versions, the number of characters was dependent on the number of keys on the
specific keyboard and the fixed total size of key maps (128). Now, the key map
size is not fixed anymore and depends on keyboard models.