Communication with keyboard via control requests.
We use "vender-specific control requests" to talk to the keyboards.

Currently, the hardware information can be read from the keyboard (number of
keys etc.). Key maps can be read, too. More requests are defined in
usbrequests.h already, but not yet implemented in the firmware.

The command line tool displays keyboard information when run without arguments.
All key maps can be downloaded from the keyboard and written to files.