The Python script can write binary key map files.
Option -M can be used to produce a key map which can be flashed directly to a
keyboard using kbuptool. Thus, it is much easier to make user-defined key maps
now. All options, in particular -m and -p, continue to work together with -M.

In principal, the output generated by -M is the same as the output of -h
compiled with avr-gcc, but we don't need a compiler anymore.