Key map selection is software-only now.
As before, key maps can be selected using Scroll Lock followed by a digit key,
but the selection is stored in EEPROM now. Thus, the key map setting is
persistently stored and survives resets and power-offs. As a result, the config
jumpers are not "wasted" for key map selection anymore and can be used for
something else.

Now, one of the jumpers is used to lock the key map to index 0, which is
primarily useful to avoid bricking the keyboard in case broken key map data was
uploaded to the keyboard and activated there.