Pressing Scroll Lock in command mode emits SL.
When pressing Scroll Lock in command mode, the keyboard returns to normal mode
as before, but also emits a Scroll Lock key stroke. Thus, Scroll Lock can be
seen by KVM switches (hit SL four times if it needs to see SL twice), and also
by the host for the few programs that actually react to this key.