Rewritten LED-enabled 122-key Model M firmware.
The firmware for the 1390702 with LED support has been rewritten to
work with the new schematic (commit b932b23). It has been tested on
breadboard and works fine.

Code for PISO shift registers has been added to shiftreg.c.

The code in scanrows.c had to be extended so to allow for alternative
ways for reading the column lines (i.e., not reading the values
directly from the I/O pins).