Any mix of re-maps can be handled by procmatrix.py.
The procmatrix.py script can combine, e.g., Dvorak or Carpalx with the
ZXCVB ergo shift map, even though certain keys get re-mapped multiple
times. In fact, any combination of re-mappings can be processed now,
even something like Colemak+Dvorak+Carpalx.

A side effect is that the order in which mappings are passed to the
script is significant now if the re-mappings are not independent of
each other. This is not a problem because exactly these cases lead to
an error in previous versions.