Implemented function key.
Any key except SL and Escape can be configured to act as function key
using the new F command. SL in function mode activates the function
key's own command mode that is used to set it's key map and to kill
the function key.

Seems to work well, but needs more testing.